+++ mrv2/docs/HISTORY.md
@@ -1,3 +1,9 @@
+v0.6.1
+------
+- Added a Scripts/Add to Script List to Python Panel.  It allows you to store
+  up to 10 scripts in the list and run them just by accessing the menu.
+  The script list is saved in the preferences.
+
 v0.6.0
 ------
 - Added the options for missing frames on the Preferences.  You can now: